+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Added `inverse_structures`
   * If true, the condition fails if the mob spawns in the listed structure
+* Added `count` to items
 * Fixed `play_sound` not working with resource pack sounds
 
 ## 4.9.1